requirements.txt
amqp==2.5.2
appdirs==1.4.3
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
distro==1.5.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
numpy==1.18.2
pandas==1.0.3
pathspec==0.7.0
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
tabula-py==2.1.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
Werkzeug==1.0.1
zipp==3.1.0
